requirements.txt
flask==3.0.3
flask-socketio==5.3.6
pycryptodome==3.20.0
simple-websocket==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1





